Faster Automatic Calibration
I have reduced the range of voltage values the automatic calibration routine tries when fitting the data. It used to check a 5000 V range (4000 V below the beam energy and 1000 V above). It now checks only a 2000 V range (from 2000 V below the beam energy to the beam energy). In addition, I have added code to automatically read in the beam energy from the header info in the new data files. The user can still enter it manually if they wish to do so.